# api/openapi.yaml
openapi: 3.1.0
info:
  title: Distributed Job Scheduler API
  version: 1.0.0
  description: >
    Submit, schedule and manage background jobs. Conventions, error codes and roles are
    described in docs/low-level-design.md §9. Every request needs
    `Authorization: Bearer <api key>`.
servers:
  - url: http://localhost:8080
security:
  - apiKey: []
paths:
  /v1/job-types:
    post:
      summary: Register a job type (admin)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/JobTypeRequest" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/JobType" } } } }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    get:
      summary: List job types
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/JobTypeList" } } } }
  /v1/job-types/{name}:
    parameters: [{ name: name, in: path, required: true, schema: { type: string } }]
    get:
      summary: Get a job type
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/JobType" } } } }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      summary: Update a job type (admin); omitted fields are unchanged
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/JobTypeRequest" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/JobType" } } } }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/job-types/{name}/pause:
    parameters: [{ name: name, in: path, required: true, schema: { type: string } }]
    post: { summary: Hold dispatch of a job type (operator); submissions and schedules continue, responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/JobType" } } } }, "404": { $ref: "#/components/responses/Error" } } }
  /v1/job-types/{name}/resume:
    parameters: [{ name: name, in: path, required: true, schema: { type: string } }]
    post: { summary: Resume dispatch of a job type (operator), responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/JobType" } } } }, "404": { $ref: "#/components/responses/Error" } } }
  /v1/jobs:
    post:
      summary: Submit a job (submitter; CRITICAL priority needs operator)
      parameters:
        - { name: Idempotency-Key, in: header, required: false, schema: { type: string, maxLength: 255 } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/SubmitJobRequest" } } } }
      responses:
        "201": { description: Created (X-Submission-Outcome created), content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
        "200": { description: Replayed or deduplicated (X-Submission-Outcome), content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "429": { description: "Rate limited (rate_limited) or over the pending-jobs quota (quota_exceeded); see Retry-After", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "503": { description: "Shed under overload (overloaded): LOW first, then NORMAL; see Retry-After", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    get:
      summary: List jobs, newest first
      parameters:
        - { name: state, in: query, schema: { $ref: "#/components/schemas/JobState" } }
        - { name: type, in: query, schema: { type: string } }
        - { name: schedule_id, in: query, schema: { type: string, format: uuid } }
        - { name: label, in: query, description: "key:value", schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/JobList" } } } }
        "422": { $ref: "#/components/responses/Error" }
  /v1/jobs/{id}:
    parameters: [{ $ref: "#/components/parameters/JobID" }]
    get:
      summary: Get a job
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/jobs/{id}/attempts:
    parameters: [{ $ref: "#/components/parameters/JobID" }]
    get:
      summary: List a job's attempts, including the running one
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AttemptList" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/jobs/{id}/cancel:
    parameters: [{ $ref: "#/components/parameters/JobID" }]
    post: { summary: Cancel a job (submitter), responses: { "200": { $ref: "#/components/responses/Job" }, "404": { $ref: "#/components/responses/Error" }, "409": { $ref: "#/components/responses/Conflict" } } }
  /v1/jobs/{id}/pause:
    parameters: [{ $ref: "#/components/parameters/JobID" }]
    post: { summary: Pause a job that hasn't started (operator), responses: { "200": { $ref: "#/components/responses/Job" }, "404": { $ref: "#/components/responses/Error" }, "409": { $ref: "#/components/responses/Conflict" } } }
  /v1/jobs/{id}/resume:
    parameters: [{ $ref: "#/components/parameters/JobID" }]
    post: { summary: Resume a paused job (operator), responses: { "200": { $ref: "#/components/responses/Job" }, "404": { $ref: "#/components/responses/Error" }, "409": { $ref: "#/components/responses/Conflict" } } }
  /v1/jobs/{id}/run:
    parameters: [{ $ref: "#/components/parameters/JobID" }]
    post: { summary: Run a delayed or paused job now (operator), responses: { "200": { $ref: "#/components/responses/Job" }, "404": { $ref: "#/components/responses/Error" }, "409": { $ref: "#/components/responses/Conflict" } } }
  /v1/jobs/{id}/retry:
    parameters: [{ $ref: "#/components/parameters/JobID" }]
    post: { summary: Retry a FAILED or DEAD_LETTERED job (operator), responses: { "200": { $ref: "#/components/responses/Job" }, "404": { $ref: "#/components/responses/Error" }, "409": { $ref: "#/components/responses/Conflict" } } }
  /v1/schedules:
    post:
      summary: Create a schedule (operator)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ScheduleRequest" } } } }
      responses:
        "201": { $ref: "#/components/responses/Schedule" }
        "409": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    get:
      summary: List schedules, newest first
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ScheduleList" } } } }
        "422": { $ref: "#/components/responses/Error" }
  /v1/schedules/{id}:
    parameters: [{ $ref: "#/components/parameters/ScheduleID" }]
    get:
      summary: Get a schedule, with its next five fire times
      responses:
        "200": { $ref: "#/components/responses/Schedule" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      summary: Update a schedule (operator); omitted fields are unchanged, null clears start_at, end_at, max_runs and priority
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ScheduleRequest" } } } }
      responses:
        "200": { $ref: "#/components/responses/Schedule" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      summary: Delete a schedule (operator); jobs that already ran stay in history
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
  /v1/schedules/{id}/pause:
    parameters: [{ $ref: "#/components/parameters/ScheduleID" }]
    post: { summary: Pause an active schedule (operator), responses: { "200": { $ref: "#/components/responses/Schedule" }, "404": { $ref: "#/components/responses/Error" }, "409": { $ref: "#/components/responses/Error" } } }
  /v1/schedules/{id}/resume:
    parameters: [{ $ref: "#/components/parameters/ScheduleID" }]
    post: { summary: Resume a paused schedule (operator); missed fires follow the misfire policy, responses: { "200": { $ref: "#/components/responses/Schedule" }, "404": { $ref: "#/components/responses/Error" }, "409": { $ref: "#/components/responses/Error" } } }
  /v1/schedules/{id}/trigger:
    parameters: [{ $ref: "#/components/parameters/ScheduleID" }]
    post:
      summary: Run a schedule once now (operator); the overlap policy applies, and the cursor and max_runs are unaffected
      responses:
        "201": { description: The job created for this run, content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } } }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /v1/workers:
    get:
      summary: List active worker sessions (platform-admin)
      parameters:
        - { name: pool, in: query, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/WorkerList" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /v1/workers/{id}/drain:
    parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
    post:
      summary: Ask a worker to finish its running jobs and deregister (platform-admin)
      responses:
        "202": { description: The worker is told on its next heartbeat or poll }
        "404": { $ref: "#/components/responses/Error" }
  /v1/workers/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
    delete:
      summary: Close a worker's session now (platform-admin); its running attempts are lost and retried
      responses:
        "204": { description: Deregistered }
        "404": { $ref: "#/components/responses/Error" }
...
  /v1/pools/{name}/pause:
    parameters: [{ $ref: "#/components/parameters/PoolName" }]
    post: { summary: Hold dispatch for a pool (platform-admin); jobs stay READY, responses: { "200": { $ref: "#/components/responses/Pool" } } }
  /v1/pools/{name}/resume:
    parameters: [{ $ref: "#/components/parameters/PoolName" }]
    post: { summary: Resume dispatch for a pool (platform-admin), responses: { "200": { $ref: "#/components/responses/Pool" } } }
...
  /v1/tenants/{id}/quotas:
    parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
    get:
      summary: A tenant's configured quotas; null means the platform default (platform-admin)
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Quotas" } } } }
        "404": { $ref: "#/components/responses/Error" }
    put:
      summary: Replace a tenant's quotas (platform-admin); omitted or null fields use the platform default
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/Quotas" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Quotas" } } } }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/operations:
    post:
      summary: Start a bulk cancel or re-drive (operator); it covers jobs created before it and runs asynchronously
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/OperationRequest" } } } }
      responses:
        "202": { description: Accepted (Location points at the operation), content: { application/json: { schema: { $ref: "#/components/schemas/Operation" } } } }
        "422": { $ref: "#/components/responses/Error" }
  /v1/operations/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
    get:
      summary: Get a bulk operation's progress
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Operation" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/api-keys:
    post:
      summary: Create an API key (admin); the key is returned only once
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/APIKeyRequest" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/APIKey" } } } }
        "422": { $ref: "#/components/responses/Error" }
  /v1/api-keys/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
    delete:
      summary: Revoke an API key (admin)
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    apiKey: { type: http, scheme: bearer, description: "jsk_<prefix>_<secret>" }
  parameters:
    JobID: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    ScheduleID: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    PoolName: { name: name, in: path, required: true, schema: { type: string } }
  responses:
    Error:
      description: Error (see LLD §9.2 for codes)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Job:
      description: The job after the operation
      content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } }
    Schedule:
      description: The schedule
      content: { application/json: { schema: { $ref: "#/components/schemas/Schedule" } } }
    Pool:
      description: The pool after the operation
      content: { application/json: { schema: { $ref: "#/components/schemas/Pool" } } }
    Conflict:
      description: Not allowed from the job's current state (details.state holds it)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, examples: [invalid_argument, not_found, conflict, rate_limited] }
            message: { type: string }
            request_id: { type: string }
            details: { type: object }
    Priority: { type: string, enum: [CRITICAL, HIGH, NORMAL, LOW] }
    JobState:
      type: string
      enum: [SCHEDULED, READY, RUNNING, RETRY_PENDING, PAUSED, SUCCEEDED, FAILED, DEAD_LETTERED, CANCELLED, EXPIRED, SKIPPED]
    Duration: { type: string, description: Go duration, examples: ["30s", "5m", "1h"] }
    RetryPolicy:
      type: object
      properties:
        strategy: { type: string, enum: [fixed, exponential] }
        initial_delay: { $ref: "#/components/schemas/Duration" }
        multiplier: { type: number }
        max_delay: { $ref: "#/components/schemas/Duration" }
        jitter: { type: string, enum: [none, full] }
        max_attempts: { type: integer }
        max_retry_duration: { $ref: "#/components/schemas/Duration" }
        max_lost_attempts: { type: integer }
    JobTypeRequest:
      type: object
      properties:
        name: { type: string, pattern: "^[a-z0-9][a-z0-9._-]{0,99}$" }
        pool: { type: string }
        default_priority: { $ref: "#/components/schemas/Priority" }
        attempt_timeout: { $ref: "#/components/schemas/Duration" }
        retry_policy: { $ref: "#/components/schemas/RetryPolicy" }
        at_most_once: { type: boolean }
        enabled: { type: boolean }
    JobType:
      type: object
      properties:
        name: { type: string }
        version: { type: integer }
        pool: { type: string }
        default_priority: { $ref: "#/components/schemas/Priority" }
        attempt_timeout: { $ref: "#/components/schemas/Duration" }
        retry_policy: { $ref: "#/components/schemas/RetryPolicy" }
        at_most_once: { type: boolean }
        enabled: { type: boolean, description: false rejects submissions and holds schedules }
        paused: { type: boolean, description: true holds dispatch; submissions and schedules continue }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    JobTypeList:
      type: object
      properties: { items: { type: array, items: { $ref: "#/components/schemas/JobType" } } }
    SubmitJobRequest:
      type: object
      required: [type]
      properties:
        type: { type: string }
        payload: { description: "Any JSON value up to 64 KiB; defaults to {}" }
        run_at: { type: string, format: date-time }
        delay: { $ref: "#/components/schemas/Duration" }
        priority: { $ref: "#/components/schemas/Priority" }
        labels: { type: object, additionalProperties: { type: string }, maxProperties: 16 }
        dedupe_key: { type: string, maxLength: 255 }
        correlation_id: { type: string, maxLength: 128 }
        start_deadline: { type: string, format: date-time }
        deadline: { type: string, format: date-time }
        attempt_timeout: { $ref: "#/components/schemas/Duration" }
        retry_policy: { $ref: "#/components/schemas/RetryPolicy" }
    Attempt:
      type: object
      properties:
        id: { type: string, format: uuid }
        number: { type: integer, description: Also the attempt's fencing token }
        state: { type: string, enum: [RUNNING, SUCCEEDED, FAILED, CANCELLED, TIMED_OUT, LOST] }
        started_at: { type: string, format: date-time }
        deadline: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        error: { type: string }
        retryable: { type: boolean }
    AttemptList:
      type: object
      properties: { items: { type: array, items: { $ref: "#/components/schemas/Attempt" } } }
    Job:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string }
        type_version: { type: integer }
        pool: { type: string }
        schedule_id: { type: string, format: uuid }
        fire_time: { type: string, format: date-time }
        state: { $ref: "#/components/schemas/JobState" }
        priority: { $ref: "#/components/schemas/Priority" }
        payload: {}
        labels: { type: object, additionalProperties: { type: string } }
        dedupe_key: { type: string }
        correlation_id: { type: string }
        created_by: { type: string }
        run_at: { type: string, format: date-time }
        start_deadline: { type: string, format: date-time }
        deadline: { type: string, format: date-time }
        attempt_timeout: { $ref: "#/components/schemas/Duration" }
        retry_policy: { $ref: "#/components/schemas/RetryPolicy" }
        attempt_count: { type: integer }
        current_attempt: { $ref: "#/components/schemas/Attempt" }
        cancel_requested_at: { type: string, format: date-time }
        last_error: { type: string }
        result: {}
        reason: { type: string }
        created_at: { type: string, format: date-time }
        ready_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
    JobList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Job" } }
        next_cursor: { type: string }
    Trigger:
      type: object
      required: [kind]
      description: "Cron (5 or 6 fields, or a macro such as @daily) in an IANA time zone, or a fixed-rate or fixed-delay interval (LLD §10.2)"
      properties:
        kind: { type: string, enum: [cron, fixed_rate, fixed_delay] }
        cron: { type: string, examples: ["0 2 * * *", "*/15 * * * *", "@hourly"] }
        time_zone: { type: string, default: UTC, examples: [Europe/Paris] }
        interval: { $ref: "#/components/schemas/Duration" }
    ScheduleRequest:
      type: object
      description: name, job_type and trigger are required on create; job_type cannot change
      properties:
        name: { type: string, pattern: "^[a-z0-9][a-z0-9._-]{0,99}$" }
        job_type: { type: string }
        payload: { description: "Any JSON value up to 64 KiB; defaults to {}" }
        labels: { type: object, additionalProperties: { type: string }, maxProperties: 16 }
        priority: { oneOf: [{ $ref: "#/components/schemas/Priority" }, { type: "null" }], description: "null uses the job type's default" }
        trigger: { $ref: "#/components/schemas/Trigger" }
        start_at: { type: [string, "null"], format: date-time }
        end_at: { type: [string, "null"], format: date-time }
        max_runs: { type: [integer, "null"], minimum: 1 }
        jitter: { $ref: "#/components/schemas/Duration" }
        misfire_policy: { type: string, enum: [fire_once, skip, fire_all], default: fire_once }
        overlap_policy: { type: string, enum: [skip, buffer_one, allow, cancel_previous], default: skip }
    Schedule:
      type: object
      description: "Jobs a schedule has materialized but that aren't due yet are provisional: pausing, editing or deleting the schedule withdraws them (ADR-016)."
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        job_type: { type: string }
        payload: {}
        labels: { type: object, additionalProperties: { type: string } }
        priority: { oneOf: [{ $ref: "#/components/schemas/Priority" }, { type: "null" }] }
        trigger: { $ref: "#/components/schemas/Trigger" }
        start_at: { type: [string, "null"], format: date-time }
        end_at: { type: [string, "null"], format: date-time }
        max_runs: { type: [integer, "null"] }
        jitter: { $ref: "#/components/schemas/Duration" }
        misfire_policy: { type: string }
        overlap_policy: { type: string }
        state: { type: string, enum: [ACTIVE, PAUSED, COMPLETED] }
        next_fire_at: { type: [string, "null"], format: date-time }
        last_fire_at: { type: [string, "null"], format: date-time }
        fire_count: { type: integer }
        upcoming: { type: array, items: { type: string, format: date-time }, description: Next five fire times (single reads only) }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ScheduleList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Schedule" } }
        next_cursor: { type: string }
...
    WorkerList:
      type: object
      properties: { items: { type: array, items: { $ref: "#/components/schemas/Worker" } } }
    Pool:
      type: object
      properties:
        name: { type: string }
        paused: { type: boolean }
        owner:
          type: [object, "null"]
          description: The engine node holding the pool's lease, if any
          properties:
            node_id: { type: string }
            address: { type: string }
            epoch: { type: integer }
            lease_expires_at: { type: string, format: date-time }
        workers: { type: integer }
        slots: { type: integer }
        ready_jobs: { type: integer }
        running_jobs: { type: integer }
...
    Quotas:
      type: object
      properties:
        rate_limit: { type: [number, "null"], description: Submissions per second across all api nodes }
        max_pending: { type: [integer, "null"], maximum: 1000000, description: Unfinished jobs }
        max_running: { type: [integer, "null"], description: Running jobs in each pool }
        max_schedules: { type: [integer, "null"], description: Active and paused schedules }
        min_schedule_interval: { type: [string, "null"], description: Go duration between 1s and 24h }
        max_payload_bytes: { type: [integer, "null"], maximum: 65536 }
...
    OperationRequest:
      type: object
      required: [kind]
      description: "cancel acts on active jobs; redrive needs filter.state FAILED or DEAD_LETTERED"
      properties:
        kind: { type: string, enum: [cancel, redrive] }
        filter: { $ref: "#/components/schemas/OperationFilter" }
    Operation:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [cancel, redrive] }
        filter: { $ref: "#/components/schemas/OperationFilter" }
        state: { type: string, enum: [PENDING, RUNNING, SUCCEEDED, FAILED] }
        succeeded: { type: integer }
        skipped: { type: integer, description: Jobs no longer eligible when reached }
        failed: { type: integer }
        last_error: { type: string }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
    APIKeyRequest:
      type: object
      required: [name, role]
      properties:
        name: { type: string, maxLength: 100 }
        role: { type: string, enum: [viewer, submitter, operator, admin, platform-admin], description: "At most the caller's own role" }
        expires_at: { type: string, format: date-time }
    APIKey:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        role: { type: string }
        key: { type: string, description: Shown only in the creation response }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }